# docs/openapi.yaml
openapi: 3.1.0
info:
  title: PenFlow API
  description: Core backend API for the PenFlow platform.
  version: 1.0.0
paths:
...
  /api/v1/engagements/:
    post:
      tags:
      - Engagements
      summary: Create client engagement request
      operationId: create_engagement_request_api_v1_engagements__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EngagementCreateRequest'
        required: true
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EngagementCreateResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/v1/engagements:
    get:
      tags:
      - Engagements
      summary: List engagements for user
      operationId: get_engagements_api_v1_engagements_get
      security:
      - HTTPBearer: []
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/EngagementStatus'
          - type: 'null'
          title: Status
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 255
          - type: 'null'
          title: Search
      - name: sort
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/EngagementSortField'
          default: updated_at
      - name: order
        in: query
        required: false
        schema:
          $ref: '#/components/schemas/SortOrder'
          default: desc
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EngagementListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/engagements/{engagement_id}:
    get:
      tags:
      - Engagements
      summary: Get engagement detail
      operationId: get_engagement_api_v1_engagements__engagement_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: engagement_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Engagement Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EngagementDetailResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/engagements/{engagement_id}/findings:
    get:
      tags:
      - Engagements
      summary: List findings for an engagement
      operationId: get_engagement_findings_api_v1_engagements__engagement_id__findings_get
      security:
      - HTTPBearer: []
      parameters:
      - name: engagement_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Engagement Id
      - name: source
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 100
          - type: 'null'
          title: Source
      - name: severity
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/Severity'
          - type: 'null'
          title: Severity
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/FindingStatus'
          - type: 'null'
          title: Status
      - name: review_status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/FindingReviewStatus'
          - type: 'null'
          title: Review Status
      - name: search
        in: query
        required: false
        schema:
          anyOf:
          - type: string
            maxLength: 255
          - type: 'null'
          title: Search
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 100
          minimum: 1
          default: 20
          title: Limit
      - name: offset
        in: query
        required: false
        schema:
          type: integer
          minimum: 0
          default: 0
          title: Offset
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/app__schemas__finding__FindingListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Engagements
      summary: Create manual finding for an engagement
      operationId: create_engagement_finding_api_v1_engagements__engagement_id__findings_post
      security:
      - HTTPBearer: []
      parameters:
      - name: engagement_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Engagement Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FindingCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FindingListItem'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/engagements/{engagement_id}/retests:
    get:
      tags:
      - Engagements
      summary: List retests for an engagement
      operationId: get_engagement_retests_api_v1_engagements__engagement_id__retests_get
      security:
      - HTTPBearer: []
      parameters:
      - name: engagement_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Engagement Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RetestListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/engagements/{engagement_id}/messages:
    get:
      tags:
      - Engagements
      summary: List engagement messages
      operationId: get_engagement_messages_api_v1_engagements__engagement_id__messages_get
      security:
      - HTTPBearer: []
      parameters:
      - name: engagement_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Engagement Id
      - name: channel
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/EngagementMessageChannel'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EngagementMessageListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - Engagements
      summary: Create engagement message
      operationId: create_engagement_message_api_v1_engagements__engagement_id__messages_post
      security:
      - HTTPBearer: []
      parameters:
      - name: engagement_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Engagement Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EngagementMessageCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EngagementMessageResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/engagements/{engagement_id}/messages/read:
    patch:
      tags:
      - Engagements
      summary: Marks viewed messages as read
      operationId: mark_engagement_messages_read_api_v1_engagements__engagement_id__messages_read_patch
      security:
      - HTTPBearer: []
      parameters:
      - name: engagement_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Engagement Id
      - name: channel
        in: query
        required: true
        schema:
          $ref: '#/components/schemas/EngagementMessageChannel'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MarkMessagesReadResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/engagements/{engagement_id}/activity:
    get:
      tags:
      - Engagements
      summary: List engagement activity
      operationId: get_engagement_activity_api_v1_engagements__engagement_id__activity_get
      security:
      - HTTPBearer: []
      parameters:
      - name: engagement_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Engagement Id
      - name: limit
        in: query
        required: false
        schema:
          type: integer
          maximum: 200
          minimum: 1
          default: 100
          title: Limit
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ActivityListResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/engagements/{engagement_id}/report:
    get:
      tags:
      - Engagements
      summary: Get engagement report
      operationId: get_engagement_report_api_v1_engagements__engagement_id__report_get
      security:
      - HTTPBearer: []
      parameters:
      - name: engagement_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Engagement Id
      - name: version
        in: query
        required: false
        schema:
          type: integer
          default: 1
          title: Version
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Get Engagement Report Api V1 Engagements  Engagement
                  Id  Report Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/v1/engagements/{engagement_id}/report/retry:
    post:
      tags:
      - Engagements
      summary: Retry engagement report generation
      operationId: retry_engagement_report_api_v1_engagements__engagement_id__report_retry_post
      security:
      - HTTPBearer: []
      parameters:
      - name: engagement_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Engagement Id
      - name: version
        in: query
        required: false
        schema:
          type: integer
          default: 1
          title: Version
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
                title: Response Retry Engagement Report Api V1 Engagements  Engagement
                  Id  Report Retry Post
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    EngagementCreateRequest:
      properties:
        engagement_type:
          $ref: '#/components/schemas/EngagementType'
        assessment_type:
          $ref: '#/components/schemas/AssessmentType'
        objective:
          type: string
          maxLength: 5000
          minLength: 1
          title: Objective
        start_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Start Date
        end_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: End Date
        constraints:
          anyOf:
          - type: string
            maxLength: 5000
          - type: 'null'
          title: Constraints
        primary_contact:
          anyOf:
          - type: string
            maxLength: 255
          - type: 'null'
          title: Primary Contact
        assets:
          items:
            $ref: '#/components/schemas/EngagementAssetCreate'
          type: array
          minItems: 1
          title: Assets
      type: object
      required:
      - engagement_type
      - assessment_type
      - objective
      - assets
      title: EngagementCreateRequest
    EngagementCreateResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        status:
          $ref: '#/components/schemas/EngagementStatus'
        engagement_type:
          $ref: '#/components/schemas/EngagementType'
        assessment_type:
          $ref: '#/components/schemas/AssessmentType'
        objective:
          type: string
          title: Objective
        start_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Start Date
        end_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: End Date
        asset_count:
          type: integer
          title: Asset Count
        estimated_quote:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Estimated Quote
        service_delivery_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Service Delivery Id
        assigned_pentester_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Assigned Pentester Id
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - id
      - status
      - engagement_type
      - assessment_type
      - objective
      - asset_count
      - estimated_quote
      - created_at
      title: EngagementCreateResponse
    EngagementDetailResponse:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        title:
          type: string
          title: Title
        engagement_type:
          $ref: '#/components/schemas/EngagementType'
        assessment_type:
          $ref: '#/components/schemas/AssessmentType'
        priority:
          type: string
          title: Priority
        status:
          $ref: '#/components/schemas/EngagementStatus'
        scope:
          type: string
          title: Scope
        estimated_quote:
          type: string
          pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          title: Estimated Quote
        final_quote:
          anyOf:
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Final Quote
        estimated_duration_days:
          anyOf:
          - type: integer
          - type: 'null'
          title: Estimated Duration Days
        requested_start_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Requested Start Date
        requested_end_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Requested End Date
        scheduled_start_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Scheduled Start Date
        scheduled_end_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Scheduled End Date
        target_date:
          anyOf:
          - type: string
            format: date
          - type: 'null'
          title: Target Date
        started_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Started At
        completed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Completed At
        reviewed_at:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Reviewed At
        review_note:
          anyOf:
          - type: string
          - type: 'null'
          title: Review Note
        created_at:
          type: string
          format: date-time
          title: Created At
        updated_at:
          type: string
          format: date-time
          title: Updated At
        client:
          $ref: '#/components/schemas/UserSummary'
        service_delivery:
          anyOf:
          - $ref: '#/components/schemas/UserSummary'
          - type: 'null'
        assigned_pentester:
          anyOf:
          - $ref: '#/components/schemas/UserSummary'
          - type: 'null'
        assets:
          items:
            $ref: '#/components/schemas/EngagementAssetResponse'
          type: array
          title: Assets
        counts:
          $ref: '#/components/schemas/EngagementOverviewCounts'
        recent_findings:
          items:
            $ref: '#/components/schemas/FindingListItem'
          type: array
          title: Recent Findings
        previous_scan:
          anyOf:
          - $ref: '#/components/schemas/PreviousScanSummary'
          - type: 'null'
      type: object
      required:
      - id
      - title
      - engagement_type
      - assessment_type
      - priority
      - status
      - scope
      - estimated_quote
      - created_at
      - updated_at
      - client
      - assets
      - counts
      - recent_findings
      title: EngagementDetailResponse
...
    EngagementListResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/EngagementListItem'
          type: array
          title: Items
        counts:
          $ref: '#/components/schemas/EngagementCounts'
        pagination:
          $ref: '#/components/schemas/EngagementPagination'
      type: object
      required:
      - items
      - counts
      - pagination
      title: EngagementListResponse
    EngagementMessageChannel:
      type: string
      enum:
      - client_service_delivery
      - service_delivery_pentester
      title: EngagementMessageChannel
...
    EngagementMessageListResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/EngagementMessageResponse'
          type: array
          title: Items
      type: object
      required:
      - items
      title: EngagementMessageListResponse
...
    EngagementSortField:
      type: string
      enum:
      - created_at
      - updated_at
      - client
      - status
      - requested_start_date
      title: EngagementSortField
    EngagementStatus:
      type: string
      enum:
      - requested
      - scoping
      - scheduled
      - in_progress
      - review
      - completed
      - cancelled
      title: EngagementStatus
...
    EngagementType:
      type: string
      enum:
      - black_box
      - grey_box
      - white_box
      title: EngagementType
...
    FindingCreate:
      properties:
        engagement_asset_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Engagement Asset Id
        title:
          type: string
          maxLength: 255
          minLength: 1
          title: Title
        severity:
          $ref: '#/components/schemas/Severity'
        cvss_score:
          anyOf:
          - type: number
            maximum: 10.0
            minimum: 0.0
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Cvss Score
        cve_id:
          anyOf:
          - type: string
            maxLength: 50
          - type: 'null'
          title: Cve Id
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        recommendation:
          anyOf:
          - type: string
          - type: 'null'
          title: Recommendation
      type: object
      required:
      - title
      - severity
      title: FindingCreate
...
    FindingListItem:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        engagement_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Engagement Id
        engagement_asset_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Engagement Asset Id
        source:
          type: string
          title: Source
        status:
          $ref: '#/components/schemas/FindingStatus'
        is_verified:
          type: boolean
          title: Is Verified
        severity:
          $ref: '#/components/schemas/Severity'
        cvss_score:
          anyOf:
          - type: string
            pattern: ^(?!^[-+.]*$)[+-]?0*\d*\.?\d*$
          - type: 'null'
          title: Cvss Score
        cve_id:
          anyOf:
          - type: string
          - type: 'null'
          title: Cve Id
        title:
          type: string
          title: Title
        description:
          anyOf:
          - type: string
          - type: 'null'
          title: Description
        created_at:
          type: string
          format: date-time
          title: Created At
        asset_identifier:
          anyOf:
          - type: string
          - type: 'null'
          title: Asset Identifier
      type: object
      required:
      - id
      - source
      - status
      - is_verified
      - severity
      - title
      - created_at
      title: FindingListItem
...
    FindingReviewStatus:
      type: string
      enum:
      - draft
      - ready_for_review
      - published
      - needs_revision
      title: FindingReviewStatus
...
    FindingStatus:
      type: string
      enum:
      - open
      - in_progress
      - resolved
      - accepted_risk
      - false_positive
      title: FindingStatus
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    MarkMessagesReadResponse:
      properties:
        marked_read:
          type: integer
          title: Marked Read
      type: object
      required:
      - marked_read
      title: MarkMessagesReadResponse
...
    RetestListResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/RetestListItem'
          type: array
          title: Items
      type: object
      required:
      - items
      title: RetestListResponse
...
    Severity:
      type: string
      enum:
      - info
      - low
      - medium
      - high
      - critical
      title: Severity
    SortOrder:
      type: string
      enum:
      - asc
      - desc
      title: SortOrder
...
    app__schemas__finding__FindingListResponse:
      properties:
        items:
          items:
            $ref: '#/components/schemas/FindingListItem'
          type: array
          title: Items
        pagination:
          $ref: '#/components/schemas/FindingPagination'
      type: object
      required:
      - items
      - pagination
      title: FindingListResponse
...
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer